+++ app/assets/javascripts/application.js
@@ -1,28 +1,3 @@
-//= require jquery
-//= require jquery.turbolinks
-//= require jquery_ujs
-//= require turbolinks
 //= require bootstrap/dropdown
 //=my home
 //=require_tree
-
-$(document).ready(function(){
-	$(".play-trailer").click(function(){
-		toggleVideo('show');
-		$(".moviecard").addClass("movie-view-trailer");
-	});
-
-	$(".back-btn").click(function(){
-		$(".moviecard").removeClass("movie-view-trailer");
-		toggleVideo('hide');
-	});	
-});
-
-function toggleVideo(state) {
-    // if state == 'hide', hide. Else: show video
-    var div = document.getElementById("youvideo");
-    var iframe = div.getElementsByTagName("iframe")[0].contentWindow;
-    div.style.display = state == 'hide' ? 'none' : '';
-    func = state == 'hide' ? 'pauseVideo' : 'playVideo';
-    iframe.postMessage('{"event":"command","func":"' + func + '","args":""}', '*');
-}
