+++ app/assets/stylesheets/application.css.scss
@@ -9,8 +9,10 @@
  * compiled file so the styles you add here take precedence over styles defined in any other CSS/SCSS
  * files in this directory. Styles in this file should be added after the last require_* statement.
  * It is generally better to create a new file per style scope.
- */
 
+*= require_tree
+*= require_self
+ */
 .alert-alert {
   @extend .alert-warning;
 }
