+++ app/views/layouts/application.html.erb
@@ -5,51 +5,52 @@
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
 
-    <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
-    <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.js"></script>
+    <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': false %>
     <%= stylesheet_link_tag 'application', media: 'all' %>
     <%= javascript_include_tag Ckeditor.cdn_url %>
-  <%= javascript_include_tag 'application' %>
-
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+    <%= javascript_include_tag 'application' %>
 
   
-  <link href="//netdna.bootstrapcdn.com/bootstrap/3.0.0/css/bootstrap.min.css" rel="stylesheet" id="bootstrap-css">
-<script src="//netdna.bootstrapcdn.com/bootstrap/3.0.0/js/bootstrap.min.js"></script>
-<script src="//code.jquery.com/jquery-1.11.1.min.js"></script>
+    <link href="//netdna.bootstrapcdn.com/bootstrap/3.0.0/css/bootstrap.min.css" rel="stylesheet" id="bootstrap-css">
+    <script src="//netdna.bootstrapcdn.com/bootstrap/3.0.0/js/bootstrap.min.js"></script>
   </head>
 
   <body>
     <nav class="navbar navbar-inverse">
-  <div class="container">
-    <div class="navbar-header">
-      <%= link_to 'IMDB', root_path, class: 'navbar-brand' %>
-    </div>
-    <div id="navbar">
-      <ul class="nav navbar-nav">
-        <li><%= link_to 'Home', root_path %></li>
-        <% if user_signed_in? %>
-       <li><%= link_to 'Secret', secret_path %></li>
-       <li><%= link_to 'Home2', home_path %></li>
-      <% end %>
-      </ul>
-      <ul class="nav navbar-nav pull-right">
-        <% if user_signed_in? %>
-          <li class="dropdown">
-
-            <a class="dropdown-toggle" data-toggle="dropdown" href="#">
-            <%= image_tag user_avatar(current_user,35), class: "img-circle" %>
+      <div class="container">
+        <div class="navbar-header">
+          <%= link_to 'IMDB', root_path, class: 'navbar-brand' %>
+        </div>
+      <div id="navbar">
+        
+        <ul class="nav navbar-nav">
+          <li><%= link_to 'Home', root_path %></li>
+          <% if user_signed_in? %>
+           <li><%= link_to 'Secret', secret_path %></li>
+           <li><%= link_to 'Home2', home_path %></li>
+          <% end %>
+        </ul>
+        
+        <ul class="nav navbar-nav pull-right">
+          <% if user_signed_in? %>
+          
+            <li class="dropdown">
+              <a class="dropdown-toggle" data-toggle="dropdown" href="#">
+              <%= image_tag user_avatar(current_user,35), class: "img-circle" %>
               <span class="caret"></span>
-            </a>
+              </a>
+          
             <ul class="dropdown-menu" role="menu">
               <li><%= link_to 'Profile', edit_user_registration_path %></li>
               <li><%= link_to 'Log out', destroy_user_session_path, method: :delete %></li>
             </ul>
           </li>
-        <% else %>
-          <li><%= link_to 'Log In', new_user_session_path %></li>
-          <li><%= link_to 'Sign Up', new_user_registration_path %></li>
-        <% end %>
+          
+          <% else %>
+            <li><%= link_to 'Log In', new_user_session_path %></li>
+            <li><%= link_to 'Sign Up', new_user_registration_path %></li>
+          <% end %>
       </ul>
     </div>
   </div>
