+++ app/views/movies/_form.html.erb
@@ -18,9 +18,7 @@
 
   <div class="form-group">
     <%= form.label :description %>
-<!--     <%= form.text_area :description %> -->
-<%= form.cktext_area :description, ckeditor: { language: 'uk'} %>
-
+    <%= form.cktext_area :description, ckeditor: { language: 'uk'} %>
   </div>
 
   <div class="form-group">
@@ -28,24 +26,45 @@
     <%= form.select :genre, ['Comedy', 'Action', 'Horror', 'Thriller', 'Romantic']%>
   </div>
 
-<div class="form-group">
-  <%= form.label :release_date %>
-  <%= form.text_field :release_date %>
-</div>
+  <div class="form-group">
+    <%= form.label :release_date %>
+    <%= form.text_field :release_date %>
+  </div>
 
-<div class="form-group">
-  <%= form.label :trailer %>
-  <%= form.text_field :trailer %>
-</div>
-  
-<div class="form-group">
-  <%= form.label :rating %>
-  <%= form.text_field :rating %>
-</div>
+  <div class="form-group">
+    <%= form.label :trailer %>
+    <%= form.text_field :trailer %>
+  </div>
 
-<div class="form-group"> 
-<%= form.file_field :images, multiple: true %> 
-</div>
+  <div class="field">
+    <%= form.label :rating %>
+    <%= form.text_field :rating %>
+  </div>
+
+  <div class="field">
+    <%= form.label :director %>
+    <%= form.text_field :director %>
+  </div>
+
+  <div class="field">
+    <%= form.label :writer %>
+    <%= form.text_field :writer %>
+  </div>
+
+  <div class="field">
+    <%= form.label :run_time %>
+    <%= form.text_field :run_time %>
+  </div>
+
+  <div class="field"> 
+    <%= form.file_field :images, multiple: true %> 
+  </div>
+
+  <div class="field">
+   <%= form.label :actors %>
+   <%= form.collection_select :actors, @actors, :id, :name, {} %>
+  </div>
+
   <div class="actions">
     <%= form.submit :class => 'btn btn-primary' %>
   </div>
