+++ app/views/movies/show.html.erb
@@ -2,31 +2,25 @@
   <div class="movie-poster play-trailer" style="background: url('<%= url_for @movie.images.first %>');background-size: contain;">
   </div>
   <div id="movie-content">
-   
-    <div class="movie-ratings"><span class="star">★</span><span class="score"> <%= @movie.rating %>  </span><span class="score-out-of">/ 10 (IMDB)</span></div>
-   
-    <div class="movie-title"><a href="http://www.imdb.com/title/tt3397884" target="_blank"><%= @movie.title %> </a><span class="movie-year"> <%= @movie.release_date %>  </span></div>
+    <div class="movie-ratings">
+      <span class="star">★</span>
+      <span class="score"> <%= @movie.rating %>  </span>
+      <span class="score-out-of">/ 10 (IMDB)</span>
+    </div>
+    
+    <div class="movie-title"><%= @movie.title %> </a><span class="movie-year"> <%= @movie.release_date %>  </span></div>
    
     <div class="movie-details"><span class="movie-rating">R</span><span class="movie-duration">2h 1min</span><span class="movie-genre"> <%= @movie.genre %> </span></div>
+    <div class="movie-castcrew"><span class="title">Director</span><span class="name"> <%= @movie.director %> </span></div>
    
-<!--     <div class="movie-castcrew"><span class="title">Director</span><span class="name">Denis Villeneuve</span></div> -->
-   
-   <!--  <div class="movie-castcrew"><span class="title">Writer</span><span class="name">Taylor Sheridan</span></div>
-    -->
-<!--     <div class="movie-castcrew"><span class="title">Cast</span><span class="name">Emily Blunt, Josh Brolin, Benicio Del Toro</span></div> -->
-   
+    <div class="movie-castcrew"><span class="title">Writer</span><span class="name"> <%= @movie.writer %> </span></div>
+    <div class="movie-castcrew"><span class="title">Actors</span></div>
     <div class="movie-synopsis"> <%=raw @movie.description %> </div>
-   
     <button class="movie-trailer-btn play-trailer" type="button">play trailer</button>
   </div>
   <div class="movie-trailer"><a class="back-btn">back</a>
     <div id="youvideo">
-
-      <!-- <iframe src= "<% @movie.trailer %>" frameborder="0" allowfullscreen="allowfullscreen" allowScriptAccess="always"></iframe> -->
-<!--       <%= video_tag( @movie.trailer ) %>
- -->
-<iframe width="490" height="275" src="<%= @movie.trailer %>" frameborder="0" allowfullscreen></iframe>
-
+      <iframe width="490" height="275" src="<%= @movie.trailer %>" frameborder="0" allowfullscreen></iframe>
     </div>
   </div>
 </div>
