+++ app/views/pages/index.html.erb
@@ -5,8 +5,7 @@
       <p>Register to get started.</p>
       <p>
         <%= link_to 'Register', new_user_registration_path, class: 'btn btn-primary btn-lg' %>
-        <%= link_to 'Movies', movies_path, class: 'btn btn-primary btn-lg' %>
-        
+        <%= link_to 'Movies', movies_path, class: 'btn btn-primary btn-lg' %>   
       </p>
     </div>
   </div>
